requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
boolean.py==4.0
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.3
cyclonedx-python-lib==7.6.2
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
emoji==2.14.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kubernetes==31.0.0
langchain==0.3.7
langchain-chroma==0.1.4
langchain-community==0.3.0
langchain-core==0.3.15
langchain-google-genai==2.0.4
langchain-google-vertexai==2.0.7
langchain-text-splitters==0.3.2
langdetect==1.0.9
langserve==0.3.0
langsmith==0.1.140
license-expression==30.4.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
marshmallow==3.23.1
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.0
opentelemetry-api==1.28.0
opentelemetry-exporter-otlp-proto-common==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-instrumentation==0.49b0
opentelemetry-instrumentation-asgi==0.49b0
opentelemetry-instrumentation-fastapi==0.49b0
opentelemetry-proto==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-semantic-conventions==0.49b0
opentelemetry-util-http==0.49b0
orjson==3.10.11
overrides==7.7.0
packageurl-python==0.16.0
packaging==24.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
posthog==3.7.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
py-serializable==1.1.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sse-starlette==1.8.2
starlette==0.41.2
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.2
tqdm==4.67.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.14.4
unstructured-client==0.27.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2
